render objects may have predefined animations and default color and opacity in the asset file

+++ include/cage-client/graphics/renderTexture.h
@@ -9,7 +9,7 @@ namespace cage
 	public:
 		void setDebugName(const string &name);
 
-		uint32 animationDuration;
+		uint64 animationDuration;
 		bool animationLoop;
 
 		uint32 getId() const;
